% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPandoraHelper.R
\name{get_capture_id}
\alias{get_capture_id}
\title{Capture ID}
\usage{
get_capture_id(id, keep_ss_suffix = FALSE)
}
\arguments{
\item{id}{The Pandora ID to extract the Capture ID from.}

\item{keep_ss_suffix}{Whether to keep the '_ss' suffix in the Capture ID.}
}
\value{
The extracted Capture ID.
}
\description{
Extract the Capture ID from a Pandora ID.
}
\examples{
get_capture_id("ABC001.A0101.SG1.1")
get_capture_id("ABC001.A0101.SG1.1", keep_ss_suffix = TRUE)
}
