% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPandoraHelper.R
\name{get_extract_id}
\alias{get_extract_id}
\title{Extract ID}
\usage{
get_extract_id(id, keep_ss_suffix = FALSE)
}
\arguments{
\item{id}{The Pandora ID to extract the Extract ID from.}

\item{keep_ss_suffix}{Whether to keep the '_ss' suffix in the Extract ID.}
}
\value{
The extracted Extract ID.
}
\description{
Extract the Extract ID from a Pandora ID.
}
\examples{
get_extract_id("ABC001.A0101")
get_extract_id("ABC001.A0101.SG1.1", keep_ss_suffix = TRUE)
}
