% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPandoraHelper.R
\name{get_ind_id}
\alias{get_ind_id}
\title{Individual ID}
\usage{
get_ind_id(id, keep_ss_suffix = FALSE)
}
\arguments{
\item{id}{The Pandora ID to extract the Individual ID from.}

\item{keep_ss_suffix}{Whether to keep the '_ss' suffix in the Individual ID.}
}
\value{
The extracted Individual ID.
}
\description{
Extract the Individual ID from a Pandora ID.
}
\examples{
get_ind_id("ABC001")
get_ind_id("ABC001.A0101", keep_ss_suffix = TRUE)
}
