% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPandoraHelper.R
\name{get_library_id}
\alias{get_library_id}
\title{Library ID}
\usage{
get_library_id(id, keep_ss_suffix = FALSE)
}
\arguments{
\item{id}{The Pandora ID to extract the Library ID from.}

\item{keep_ss_suffix}{Whether to keep the '_ss' suffix in the Library ID.}
}
\value{
The extracted Library ID.
}
\description{
Extract the Library ID from a Pandora ID.
}
\examples{
get_library_id("ABC001.A0101")
get_library_id("ABC001.A0101.SG1.1", keep_ss_suffix = TRUE)
}
