% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPandoraHelper.R
\name{get_sample_id}
\alias{get_sample_id}
\title{Sample ID}
\usage{
get_sample_id(id, keep_ss_suffix = FALSE)
}
\arguments{
\item{id}{The Pandora ID to extract the Sample ID from.}

\item{keep_ss_suffix}{Whether to keep the '_ss' suffix in the Sample ID.}
}
\value{
The extracted Sample ID.
}
\description{
Extract the Sample ID from a Pandora ID.
}
\examples{
get_sample_id("ABC001.A0101")
get_sample_id("ABC001.A0101.SG1.1", keep_ss_suffix = TRUE)
}
