% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPandoraHelper.R
\name{get_sequencing_id}
\alias{get_sequencing_id}
\title{Sequencing ID}
\usage{
get_sequencing_id(id, keep_ss_suffix = FALSE)
}
\arguments{
\item{id}{The Pandora ID to extract the Sequencing ID from.}

\item{keep_ss_suffix}{Whether to keep the '_ss' suffix in the Sequencing ID.}
}
\value{
The extracted Sequencing ID.
}
\description{
Extract the Sequencing ID from a Pandora ID.
}
\examples{
get_sequencing_id("ABC001.A0101.SG1.1")
get_sequencing_id("ABC001.A0101.SG1.1", keep_ss_suffix = TRUE)
}
