% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPandoraHelper.R
\name{get_site_id}
\alias{get_site_id}
\title{Extract Site ID from Pandora ID}
\usage{
get_site_id(id, keep_ss_suffix = FALSE)
}
\arguments{
\item{id}{The Pandora ID to extract the Site ID from.}

\item{keep_ss_suffix}{Whether to keep the '_ss' suffix in the Site ID.}
}
\value{
The extracted Site ID.
}
\description{
Extract the Site ID from a Pandora ID.
}
\examples{
get_site_id("ABC001.A0101")
get_site_id("ABC001.A0101.SG1.1", keep_ss_suffix = TRUE)
}
