% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPandoraHelper.R
\name{remove_suffix}
\alias{remove_suffix}
\title{Helper Functions for Parsing Pandora IDs}
\usage{
remove_suffix(ind_id, keep_ss_suffix = FALSE)
}
\arguments{
\item{ind_id}{The ID to remove the suffix from.}

\item{keep_ss_suffix}{Whether to keep the '_ss' suffix in the ID.}
}
\value{
The ID with the suffix removed if keep_ss_suffix is FALSE, otherwise the original ID.
}
\description{
This package provides helper functions for parsing Pandora IDs.

Remove the '_ss' suffix from an ID if present.
}
\examples{
remove_suffix("ABC001_ss")
remove_suffix("ABC001_ss", keep_ss_suffix = TRUE)
}
\author{
Thiseas C. Lamnidis
}
